Add more functions to remove new dirs

+++ scripts/deinstall.sh
@@ -6,15 +6,43 @@ if [ "$(id -u)" -ne 0 ]; then
     exit 1
 fi
 
+remove_applications()
+{
+  if [ -d "/Applications" ] ; then
+    rm -rf /Applications
+  fi
+}
+
+remove_developer()
+{
+  if [ -d "/Developer" ] ;
+    rm -rf /Developer
+  fi
+}
+
 remove_local()
 {
-  rm -rf /Local
+  if [ -d "/Local" ] ; then
+    rm -rf /Local
+  fi
+}
+
+remove_opt()
+{
+  if [ -d "/opt" ] ; then
+    rm -rf /opt
+  fi
 }
 
 remove_system()
 {
-  rm -rf /System
+  if [ -d "/System" ] ; then
+    rm -rf /System
+  fi
 }
 
+remove_applications
+remove_developer
 remove_local
+remove_opt
 remove_system
